chore: restructure codebase
