+++ components/hord-cli/src/db/mod.rs
@@ -1,48 +1,29 @@
 use std::{
-    collections::{BTreeMap, HashMap, VecDeque},
+    collections::BTreeMap,
     hash::BuildHasherDefault,
+    io::{Read, Write},
     path::PathBuf,
-    sync::{mpsc::Sender, Arc},
-    thread::sleep,
-    time::Duration,
+    sync::Arc,
 };
 
-use chainhook_sdk::{
-    indexer::bitcoin::{
-        build_http_client, download_block_with_retry, parse_downloaded_block,
-        retrieve_block_hash_with_retry, try_download_block_bytes_with_retry,
-    },
-    types::{
-        BitcoinBlockData, BlockIdentifier, OrdinalInscriptionRevealData,
-        OrdinalInscriptionTransferData, TransactionIdentifier,
-    },
-};
-use crossbeam_channel::bounded;
 use dashmap::DashMap;
 use fxhash::FxHasher;
-use hiro_system_kit::slog;
 use rand::{thread_rng, Rng};
 
 use rocksdb::DB;
 use rusqlite::{Connection, OpenFlags, ToSql, Transaction};
 use std::io::Cursor;
-use std::io::{Read, Write};
-use threadpool::ThreadPool;
-use tokio::task::JoinSet;
 
 use chainhook_sdk::{
-    indexer::bitcoin::BitcoinBlockFullBreakdown, observer::BitcoinConfig, utils::Context,
+    indexer::bitcoin::BitcoinBlockFullBreakdown,
+    types::{
+        BitcoinBlockData, BlockIdentifier, OrdinalInscriptionRevealData,
+        OrdinalInscriptionTransferData, TransactionIdentifier,
+    },
+    utils::Context,
 };
 
-use crate::hord::{
-    new_traversals_lazy_cache, update_hord_db_and_augment_bitcoin_block,
-    update_hord_db_and_augment_bitcoin_block_v3,
-};
 use crate::ord::{height::Height, sat::Sat};
-use crate::{
-    config::Config,
-    hord::{self, HordConfig},
-};
 
 #[derive(Clone, Debug)]
 pub struct InscriptionHeigthHint {
@@ -92,7 +73,7 @@ pub fn initialize_hord_db(path: &PathBuf, ctx: &Context) -> Connection {
         [],
     ) {
         ctx.try_log(|logger| {
-            slog::warn!(
+            warn!(
                 logger,
                 "Unable to create table inscriptions: {}",
                 e.to_string()
@@ -111,7 +92,7 @@ pub fn initialize_hord_db(path: &PathBuf, ctx: &Context) -> Connection {
             [],
         ) {
             ctx.try_log(|logger| {
-                slog::warn!(logger, "Unable to create table locations:{}", e.to_string())
+                warn!(logger, "Unable to create table locations:{}", e.to_string())
             });
         }
 
@@ -119,38 +100,38 @@ pub fn initialize_hord_db(path: &PathBuf, ctx: &Context) -> Connection {
             "CREATE INDEX IF NOT EXISTS index_inscriptions_on_ordinal_number ON inscriptions(ordinal_number);",
             [],
         ) {
-            ctx.try_log(|logger| slog::warn!(logger, "{}", e.to_string()));
+            ctx.try_log(|logger| warn!(logger, "{}", e.to_string()));
         }
         if let Err(e) = conn.execute(
             "CREATE INDEX IF NOT EXISTS index_inscriptions_on_inscription_number ON inscriptions(inscription_number);",
             [],
         ) {
-            ctx.try_log(|logger| slog::warn!(logger, "{}", e.to_string()));
+            ctx.try_log(|logger| warn!(logger, "{}", e.to_string()));
         }
 
         if let Err(e) = conn.execute(
             "CREATE INDEX IF NOT EXISTS index_inscriptions_on_block_height ON inscriptions(block_height);",
             [],
         ) {
-            ctx.try_log(|logger| slog::warn!(logger, "{}", e.to_string()));
+            ctx.try_log(|logger| warn!(logger, "{}", e.to_string()));
         }
         if let Err(e) = conn.execute(
             "CREATE INDEX IF NOT EXISTS index_locations_on_block_height ON locations(block_height);",
             [],
         ) {
-            ctx.try_log(|logger| slog::warn!(logger, "{}", e.to_string()));
+            ctx.try_log(|logger| warn!(logger, "{}", e.to_string()));
         }
         if let Err(e) = conn.execute(
             "CREATE INDEX IF NOT EXISTS index_locations_on_outpoint_to_watch ON locations(outpoint_to_watch);",
             [],
         ) {
-            ctx.try_log(|logger| slog::warn!(logger, "{}", e.to_string()));
+            ctx.try_log(|logger| warn!(logger, "{}", e.to_string()));
         }
         if let Err(e) = conn.execute(
             "CREATE INDEX IF NOT EXISTS index_locations_on_inscription_id ON locations(inscription_id);",
             [],
         ) {
-            ctx.try_log(|logger| slog::warn!(logger, "{}", e.to_string()));
+            ctx.try_log(|logger| warn!(logger, "{}", e.to_string()));
         }
     }
     conn
@@ -164,7 +145,7 @@ fn create_or_open_readwrite_db(cache_path: &PathBuf, ctx: &Context) -> Connectio
                 // need to create
                 if let Some(dirp) = PathBuf::from(&path).parent() {
                     std::fs::create_dir_all(dirp).unwrap_or_else(|e| {
-                        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+                        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
                     });
                 }
                 OpenFlags::SQLITE_OPEN_READ_WRITE | OpenFlags::SQLITE_OPEN_CREATE
@@ -182,7 +163,7 @@ fn create_or_open_readwrite_db(cache_path: &PathBuf, ctx: &Context) -> Connectio
         match Connection::open_with_flags(&path, open_flags) {
             Ok(conn) => break conn,
             Err(e) => {
-                ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+                ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
             }
         };
         std::thread::sleep(std::time::Duration::from_secs(1));
@@ -216,7 +197,7 @@ fn open_existing_readonly_db(path: &PathBuf, ctx: &Context) -> Connection {
         match Connection::open_with_flags(path, open_flags) {
             Ok(conn) => break conn,
             Err(e) => {
-                ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+                ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
             }
         };
         std::thread::sleep(std::time::Duration::from_secs(1));
@@ -267,7 +248,7 @@ pub fn open_readonly_hord_db_conn_rocks_db_loop(base_dir: &PathBuf, ctx: &Contex
             Ok(db) => break db,
             Err(e) => {
                 ctx.try_log(|logger| {
-                    slog::warn!(logger, "Unable to open db: {e}",);
+                    warn!(logger, "Unable to open db: {e}",);
                 });
                 continue;
             }
@@ -338,10 +319,9 @@ pub fn find_lazy_block_at_block_height(
             _ => {
                 if attempt == 1 && try_iterator {
                     ctx.try_log(|logger| {
-                        slog::warn!(
+                        warn!(
                             logger,
-                            "Attempt to retrieve block {} through iterator",
-                            block_height,
+                            "Attempt to retrieve block {} through iterator", block_height,
                         )
                     });
                     let mut iter = blocks_db.iterator(rocksdb::IteratorMode::End);
@@ -356,11 +336,9 @@ pub fn find_lazy_block_at_block_height(
                 backoff = 2.0 * backoff + (backoff * rng.gen_range(0.0..1.0));
                 let duration = std::time::Duration::from_millis((backoff * 1_000.0) as u64);
                 ctx.try_log(|logger| {
-                    slog::warn!(
+                    warn!(
                         logger,
-                        "Unable to find block {}, will retry in {:?}",
-                        block_height,
-                        duration
+                        "Unable to find block {}, will retry in {:?}", block_height, duration
                     )
                 });
                 std::thread::sleep(duration);
@@ -374,7 +352,7 @@ pub fn find_lazy_block_at_block_height(
 
 pub fn remove_entry_from_blocks(block_height: u32, blocks_db_rw: &DB, ctx: &Context) {
     if let Err(e) = blocks_db_rw.delete(block_height.to_be_bytes()) {
-        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
     }
 }
 
@@ -403,7 +381,7 @@ pub fn insert_entry_in_inscriptions(
         "INSERT INTO inscriptions (inscription_id, ordinal_number, inscription_number, block_height) VALUES (?1, ?2, ?3, ?4)",
         rusqlite::params![&inscription_data.inscription_id, &inscription_data.ordinal_number, &inscription_data.inscription_number, &block_identifier.index],
     ) {
-        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
     }
     insert_inscription_in_locations(
         &inscription_data,
@@ -426,7 +404,7 @@ pub fn insert_inscription_in_locations(
         "INSERT INTO locations (inscription_id, outpoint_to_watch, offset, block_height, tx_index) VALUES (?1, ?2, ?3, ?4, ?5)",
         rusqlite::params![&inscription_data.inscription_id, &outpoint_to_watch, offset, &block_identifier.index, &inscription_data.tx_index],
     ) {
-        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
     }
 }
 
@@ -442,7 +420,7 @@ pub fn insert_transfer_in_locations_tx(
         "INSERT INTO locations (inscription_id, outpoint_to_watch, offset, block_height, tx_index) VALUES (?1, ?2, ?3, ?4, ?5)",
         rusqlite::params![&transfer_data.inscription_id, &outpoint_to_watch, offset, &block_identifier.index, &transfer_data.tx_index],
     ) {
-        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
     }
 }
 
@@ -458,7 +436,7 @@ pub fn insert_transfer_in_locations(
         "INSERT INTO locations (inscription_id, outpoint_to_watch, offset, block_height, tx_index) VALUES (?1, ?2, ?3, ?4, ?5)",
         rusqlite::params![&transfer_data.inscription_id, &outpoint_to_watch, offset, &block_identifier.index, &transfer_data.tx_index],
     ) {
-        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
     }
 }
 
@@ -722,7 +700,7 @@ pub fn find_inscription_with_id(
             Ok(stmt) => break stmt,
             Err(e) => {
                 ctx.try_log(|logger| {
-                    slog::warn!(
+                    warn!(
                         logger,
                         "unable to retrieve inscription with id: {}",
                         e.to_string(),
@@ -874,13 +852,13 @@ pub fn delete_inscriptions_in_block_range(
         "DELETE FROM inscriptions WHERE block_height >= ?1 AND block_height <= ?2",
         rusqlite::params![&start_block, &end_block],
     ) {
-        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
     }
     if let Err(e) = inscriptions_db_conn_rw.execute(
         "DELETE FROM locations WHERE block_height >= ?1 AND block_height <= ?2",
         rusqlite::params![&start_block, &end_block],
     ) {
-        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
     }
 }
 
@@ -893,13 +871,13 @@ pub fn remove_entry_from_inscriptions(
         "DELETE FROM inscriptions WHERE inscription_id = ?1",
         rusqlite::params![&inscription_id],
     ) {
-        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
     }
     if let Err(e) = inscriptions_db_rw_conn.execute(
         "DELETE FROM locations WHERE inscription_id = ?1",
         rusqlite::params![&inscription_id],
     ) {
-        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
     }
 }
 
@@ -912,7 +890,7 @@ pub fn remove_entries_from_locations_at_block_height(
         "DELETE FROM locations WHERE block_height = ?1",
         rusqlite::params![&block_height],
     ) {
-        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
     }
 }
 
@@ -931,7 +909,7 @@ pub fn insert_entry_in_locations(
         "INSERT INTO locations (inscription_id, outpoint_to_watch, offset, block_height, tx_index) VALUES (?1, ?2, ?3, ?4, ?5)",
         rusqlite::params![&inscription_id, &outpoint_to_watch, &transfer_data.inscription_offset_intra_output, &block_height, &transfer_data.tx_index],
     ) {
-        ctx.try_log(|logger| slog::error!(logger, "{}", e.to_string()));
+        ctx.try_log(|logger| error!(logger, "{}", e.to_string()));
     }
 }
 
@@ -949,250 +927,6 @@ pub fn delete_data_in_hord_db(
         inscriptions_db_conn_rw,
         &ctx,
     );
-    Ok(())
-}
-
-pub async fn fetch_and_cache_blocks_in_hord_db(
-    bitcoin_config: &BitcoinConfig,
-    blocks_db_rw: &DB,
-    inscriptions_db_conn_rw: &mut Connection,
-    start_block: u64,
-    end_block: u64,
-    hord_config: &HordConfig,
-    block_post_processor: Option<Sender<BitcoinBlockData>>,
-    ctx: &Context,
-) -> Result<(), String> {
-    let ordinal_computing_height = hord_config.first_inscription_height;
-    let number_of_blocks_to_process = end_block - start_block + 1;
-    let (block_hash_req_lim, block_req_lim, block_process_lim) =
-        if start_block >= ordinal_computing_height {
-            (32, 24, 24)
-        } else {
-            (256, 128, 128)
-        };
-    let retrieve_block_hash_pool = ThreadPool::new(hord_config.network_thread_max);
-    let (block_hash_tx, block_hash_rx) = crossbeam_channel::bounded(block_hash_req_lim);
-    let retrieve_block_data_pool = ThreadPool::new(hord_config.network_thread_max);
-    let (block_data_tx, block_data_rx) = crossbeam_channel::bounded(block_req_lim);
-    let compress_block_data_pool = ThreadPool::new(hord_config.ingestion_thread_max);
-    let (block_compressed_tx, block_compressed_rx) = crossbeam_channel::bounded(block_process_lim);
-
-    // Thread pool #1: given a block height, retrieve the block hash
-    let http_client = build_http_client();
-
-    for block_cursor in start_block..=end_block {
-        let block_height = block_cursor.clone();
-        let block_hash_tx = block_hash_tx.clone();
-        let config = bitcoin_config.clone();
-        let moved_ctx = ctx.clone();
-        let moved_http_client = http_client.clone();
-
-        retrieve_block_hash_pool.execute(move || {
-            let future = retrieve_block_hash_with_retry(
-                &moved_http_client,
-                &block_height,
-                &config,
-                &moved_ctx,
-            );
-            let block_hash = hiro_system_kit::nestable_block_on(future).unwrap();
-            block_hash_tx
-                .send(Some((block_height, block_hash)))
-                .expect("unable to channel block_hash");
-        })
-    }
-
-    // Thread pool #2: given a block hash, retrieve the full block (verbosity max, including prevout)
-    let bitcoin_network = bitcoin_config.network.clone();
-    let bitcoin_config = bitcoin_config.clone();
-    let moved_ctx = ctx.clone();
-    let block_data_tx_moved = block_data_tx.clone();
-    let _ = hiro_system_kit::thread_named("Block data retrieval")
-        .spawn(move || {
-            while let Ok(Some((block_height, block_hash))) = block_hash_rx.recv() {
-                let moved_bitcoin_config = bitcoin_config.clone();
-                let block_data_tx = block_data_tx_moved.clone();
-                let moved_ctx = moved_ctx.clone();
-                let moved_http_client = http_client.clone();
-                retrieve_block_data_pool.execute(move || {
-                    moved_ctx.try_log(|logger| debug!(logger, "Fetching block #{block_height}"));
-                    let future = download_block_with_retry(
-                        &moved_http_client,
-                        &block_hash,
-                        &moved_bitcoin_config,
-                        &moved_ctx,
-                    );
-                    let res = match hiro_system_kit::nestable_block_on(future) {
-                        Ok(block_data) => Some(block_data),
-                        Err(e) => {
-                            moved_ctx.try_log(|logger| {
-                                error!(logger, "unable to fetch block #{block_height}: {e}")
-                            });
-                            None
-                        }
-                    };
-                    let _ = block_data_tx.send(res);
-                });
-                // TODO: remove this join?
-                if block_height >= ordinal_computing_height {
-                    let _ = retrieve_block_data_pool.join();
-                }
-            }
-            let res = retrieve_block_data_pool.join();
-            res
-        })
-        .expect("unable to spawn thread");
-
-    let _ = hiro_system_kit::thread_named("Block data compression")
-        .spawn(move || {
-            while let Ok(Some(block_data)) = block_data_rx.recv() {
-                let block_compressed_tx_moved = block_compressed_tx.clone();
-                let block_height = block_data.height as u64;
-                compress_block_data_pool.execute(move || {
-                    let compressed_block =
-                        LazyBlock::from_full_block(&block_data).expect("unable to serialize block");
-                    let block_index = block_data.height as u32;
-                    let _ = block_compressed_tx_moved.send(Some((
-                        block_index,
-                        compressed_block,
-                        block_data,
-                    )));
-                });
-                if block_height >= ordinal_computing_height {
-                    let _ = compress_block_data_pool.join();
-                }
-            }
-            let res = compress_block_data_pool.join();
-            res
-        })
-        .expect("unable to spawn thread");
-
-    let mut blocks_stored = 0;
-    let mut cursor = start_block as usize;
-    let mut inbox = HashMap::new();
-    let mut num_writes = 0;
-    let traversals_cache = Arc::new(new_traversals_lazy_cache(hord_config.cache_size));
-
-    while let Ok(Some((block_height, compacted_block, raw_block))) = block_compressed_rx.recv() {
-        insert_entry_in_blocks(block_height, &compacted_block, &blocks_db_rw, &ctx);
-        blocks_stored += 1;
-        num_writes += 1;
-
-        // In the context of ordinals, we're constrained to process blocks sequentially
-        // Blocks are processed by a threadpool and could be coming out of order.
-        // Inbox block for later if the current block is not the one we should be
-        // processing.
-
-        // Should we start look for inscriptions data in blocks?
-        if raw_block.height as u64 >= ordinal_computing_height {
-            if (cursor as u64) < ordinal_computing_height {
-                cursor = raw_block.height;
-            }
-            ctx.try_log(|logger| {
-                slog::info!(
-                    logger,
-                    "Queueing compacted block #{block_height} (#{cursor})",
-                )
-            });
-            // Is the action of processing a block allows us
-            // to process more blocks present in the inbox?
-            inbox.insert(raw_block.height, raw_block);
-            while let Some(next_block) = inbox.remove(&cursor) {
-                ctx.try_log(|logger| {
-                    slog::info!(
-                        logger,
-                        "Dequeuing block #{cursor} for processing (# blocks inboxed: {})",
-                        inbox.len()
-                    )
-                });
-                let mut new_block = match hord::parse_ordinals_and_standardize_block(
-                    next_block,
-                    &bitcoin_network,
-                    &ctx,
-                ) {
-                    Ok(block) => block,
-                    Err((e, _)) => {
-                        ctx.try_log(|logger| {
-                            slog::error!(logger, "Unable to standardize bitcoin block: {e}",)
-                        });
-                        return Err(e);
-                    }
-                };
-
-                let _ = blocks_db_rw.flush();
-                if let Err(e) = update_hord_db_and_augment_bitcoin_block(
-                    &mut new_block,
-                    blocks_db_rw,
-                    inscriptions_db_conn_rw,
-                    false,
-                    &hord_config,
-                    &traversals_cache,
-                    &ctx,
-                ) {
-                    ctx.try_log(|logger| {
-                        slog::error!(
-                            logger,
-                            "Unable to augment bitcoin block {} with hord_db: {e}",
-                            new_block.block_identifier.index
-                        )
-                    });
-                    return Err(e);
-                }
-
-                if let Some(ref tx) = block_post_processor {
-                    let _ = tx.send(new_block);
-                }
-                cursor += 1;
-            }
-        } else {
-            ctx.try_log(|logger| slog::info!(logger, "Storing compacted block #{block_height}",));
-        }
-
-        if blocks_stored == number_of_blocks_to_process {
-            let _ = block_data_tx.send(None);
-            let _ = block_hash_tx.send(None);
-            ctx.try_log(|logger| {
-                slog::info!(
-                    logger,
-                    "Local ordinals storage successfully seeded with #{blocks_stored} blocks"
-                )
-            });
-            return Ok(());
-        }
-
-        if !traversals_cache.is_empty() {
-            if num_writes % 24 == 0 {
-                ctx.try_log(|logger| {
-                    slog::info!(
-                        logger,
-                        "Flushing traversals cache (#{} entries)",
-                        traversals_cache.len()
-                    );
-                });
-                traversals_cache.clear();
-            }
-        }
-
-        if num_writes % 128 == 0 {
-            ctx.try_log(|logger| {
-                slog::info!(logger, "Flushing DB to disk ({num_writes} inserts)");
-            });
-            if let Err(e) = blocks_db_rw.flush() {
-                ctx.try_log(|logger| {
-                    slog::error!(logger, "{}", e.to_string());
-                });
-            }
-            num_writes = 0;
-        }
-    }
-
-    if let Err(e) = blocks_db_rw.flush() {
-        ctx.try_log(|logger| {
-            slog::error!(logger, "{}", e.to_string());
-        });
-    }
-
-    retrieve_block_hash_pool.join();
-
     Ok(())
 }
 
@@ -1342,7 +1076,7 @@ pub fn retrieve_satoshi_point_using_lazy_storage(
         }
     }
     ctx.try_log(|logger| {
-        slog::info!(
+        info!(
             logger,
             "Computing ordinal number for Satoshi point {} ({}:0 -> {}:{}/{})  (block #{})",
             transaction_identifier.hash,
@@ -1398,7 +1132,7 @@ pub fn retrieve_satoshi_point_using_lazy_storage(
 
             if sats_in == 0 {
                 ctx.try_log(|logger| {
-                    slog::error!(
+                    error!(
                         logger,
                         "Transaction {} is originating from a non spending transaction",
                         transaction_identifier.hash
@@ -1523,7 +1257,7 @@ pub fn retrieve_satoshi_point_using_lazy_storage(
 
             if sats_in == 0 {
                 ctx.try_log(|logger| {
-                    slog::error!(
+                    error!(
                         logger,
                         "Transaction {} is originating from a non spending transaction",
                         transaction_identifier.hash
@@ -1634,7 +1368,7 @@ pub fn retrieve_satoshi_point_using_lazy_storage_v3(
         }
     }
     ctx.try_log(|logger| {
-        slog::info!(
+        info!(
             logger,
             "Computing ordinal number for Satoshi point {} ({}:0 -> {}:{}/{})  (block #{})",
             transaction_identifier.hash,
@@ -1690,7 +1424,7 @@ pub fn retrieve_satoshi_point_using_lazy_storage_v3(
 
             if sats_in == 0 {
                 ctx.try_log(|logger| {
-                    slog::error!(
+                    error!(
                         logger,
                         "Transaction {} is originating from a non spending transaction",
                         transaction_identifier.hash
@@ -1815,7 +1549,7 @@ pub fn retrieve_satoshi_point_using_lazy_storage_v3(
 
             if sats_in == 0 {
                 ctx.try_log(|logger| {
-                    slog::error!(
+                    error!(
                         logger,
                         "Transaction {} is originating from a non spending transaction",
                         transaction_identifier.hash
@@ -2133,7 +1867,7 @@ impl LazyBlock {
         for coinbase_output in block.transactions[0].metadata.outputs.iter() {
             coinbase_value += coinbase_output.value;
         }
-        buffer.write(&coinbase_value.to_be_bytes())?;
+        buffer.write_all(&coinbase_value.to_be_bytes())?;
         // For each transaction:
         for tx in block.transactions.iter().skip(1) {
             // txid - 8 first bytes
@@ -2203,262 +1937,4 @@ impl<'a> Iterator for LazyBlockTransactionIterator<'a> {
             outputs_len,
         ))
     }
-}
-
-pub fn process_blocks(
-    mut next_blocks: Vec<BitcoinBlockData>,
-    cache_l2: &Arc<DashMap<(u32, [u8; 8]), LazyBlockTransaction, BuildHasherDefault<FxHasher>>>,
-    inscription_height_hint: &mut InscriptionHeigthHint,
-    inscriptions_db_conn_rw: &mut Connection,
-    hord_config: &HordConfig,
-    post_processor: &Option<Sender<BitcoinBlockData>>,
-    ctx: &Context,
-) {
-    let mut cache_l1 = HashMap::new();
-
-    for _cursor in 0..next_blocks.len() {
-        let mut block = next_blocks.remove(0);
-
-        let _ = process_block(
-            &mut block,
-            &next_blocks,
-            &mut cache_l1,
-            cache_l2,
-            inscription_height_hint,
-            inscriptions_db_conn_rw,
-            hord_config,
-            ctx,
-        );
-
-        if let Some(post_processor_tx) = post_processor {
-            let _ = post_processor_tx.send(block);
-        }
-    }
-}
-
-pub fn process_block(
-    block: &mut BitcoinBlockData,
-    next_blocks: &Vec<BitcoinBlockData>,
-    cache_l1: &mut HashMap<(TransactionIdentifier, usize), TraversalResult>,
-    cache_l2: &Arc<DashMap<(u32, [u8; 8]), LazyBlockTransaction, BuildHasherDefault<FxHasher>>>,
-    inscription_height_hint: &mut InscriptionHeigthHint,
-    inscriptions_db_conn_rw: &mut Connection,
-    hord_config: &HordConfig,
-    ctx: &Context,
-) -> Result<(), String> {
-    update_hord_db_and_augment_bitcoin_block_v3(
-        block,
-        next_blocks,
-        cache_l1,
-        cache_l2,
-        inscription_height_hint,
-        inscriptions_db_conn_rw,
-        hord_config,
-        ctx,
-    )
-}
-
-pub async fn rebuild_rocks_db(
-    config: &Config,
-    start_block: u64,
-    end_block: u64,
-    start_sequencing_blocks_at_height: u64,
-    blocks_post_processor: Option<crossbeam_channel::Sender<Vec<(BitcoinBlockData, LazyBlock)>>>,
-    ctx: &Context,
-) -> Result<(), String> {
-    // let guard = pprof::ProfilerGuardBuilder::default()
-    //     .frequency(20)
-    //     .blocklist(&["libc", "libgcc", "pthread", "vdso"])
-    //     .build()
-    //     .unwrap();
-
-    let bitcoin_config = BitcoinConfig {
-        username: config.network.bitcoind_rpc_username.clone(),
-        password: config.network.bitcoind_rpc_password.clone(),
-        rpc_url: config.network.bitcoind_rpc_url.clone(),
-        network: config.network.bitcoin_network.clone(),
-        bitcoin_block_signaling: config.network.bitcoin_block_signaling.clone(),
-    };
-
-    let hord_config = config.get_hord_config();
-
-    let number_of_blocks_to_process = end_block - start_block + 1;
-
-    let (block_compressed_tx, block_compressed_rx) = crossbeam_channel::bounded(50);
-    let http_client = build_http_client();
-
-    let moved_config = bitcoin_config.clone();
-    let moved_ctx = ctx.clone();
-    let moved_http_client = http_client.clone();
-
-    let mut set = JoinSet::new();
-
-    let mut block_heights = VecDeque::from((start_block..=end_block).collect::<Vec<u64>>());
-
-    for _ in 0..4 {
-        if let Some(block_height) = block_heights.pop_front() {
-            let config = moved_config.clone();
-            let ctx = moved_ctx.clone();
-            let http_client = moved_http_client.clone();
-            sleep(Duration::from_millis(500));
-            set.spawn(try_download_block_bytes_with_retry(
-                http_client,
-                block_height,
-                config,
-                ctx,
-            ));
-        }
-    }
-
-    let moved_ctx: Context = ctx.clone();
-    let moved_bitcoin_network = bitcoin_config.network.clone();
-
-    let mut tx_thread_pool = vec![];
-    let mut rx_thread_pool = vec![];
-    let mut thread_pool_handles = vec![];
-
-    for _ in 0..hord_config.ingestion_thread_max {
-        let (tx, rx) = bounded::<Option<Vec<u8>>>(hord_config.ingestion_thread_queue_size);
-        tx_thread_pool.push(tx);
-        rx_thread_pool.push(rx);
-    }
-
-    for rx in rx_thread_pool.into_iter() {
-        let block_compressed_tx_moved = block_compressed_tx.clone();
-        let moved_ctx: Context = moved_ctx.clone();
-        let moved_bitcoin_network = moved_bitcoin_network.clone();
-
-        let handle = hiro_system_kit::thread_named("Block data compression")
-            .spawn(move || {
-                while let Ok(Some(block_bytes)) = rx.recv() {
-                    let raw_block_data =
-                        parse_downloaded_block(block_bytes).expect("unable to parse block");
-                    let compressed_block = LazyBlock::from_full_block(&raw_block_data)
-                        .expect("unable to compress block");
-                    let block_data = hord::parse_ordinals_and_standardize_block(
-                        raw_block_data,
-                        &moved_bitcoin_network,
-                        &moved_ctx,
-                    )
-                    .expect("unable to deserialize block");
-
-                    let _ = block_compressed_tx_moved.send(Some((block_data, compressed_block)));
-                }
-            })
-            .expect("unable to spawn thread");
-        thread_pool_handles.push(handle);
-    }
-
-    let cloned_ctx = ctx.clone();
-
-    let storage_thread = hiro_system_kit::thread_named("Ordered blocks dispatcher")
-        .spawn(move || {
-            let mut inbox = HashMap::new();
-            let mut inbox_cursor = start_sequencing_blocks_at_height.max(start_block);
-            let mut blocks_processed = 0;
-
-            loop {
-                // Dequeue all the blocks available
-                let mut new_blocks = vec![];
-                while let Ok(Some((block, compacted_block))) = block_compressed_rx.try_recv()
-                {
-                    blocks_processed += 1;
-                    new_blocks.push((block, compacted_block))
-                }
-                // 
-                let mut ooo_processing = vec![];
-                for (block, compacted_block) in new_blocks.into_iter() {
-                    let block_index = block.block_identifier.index;
-                    if block_index >= start_sequencing_blocks_at_height {
-                        inbox.insert(block_index, (block, compacted_block));
-                    } else {
-                        ooo_processing.push((block, compacted_block));
-                        // todo: do something
-                    }
-                }
-
-                // In order processing: construct the longest sequence of known blocks
-                let mut chunk = Vec::new();
-                while let Some((block, compacted_block)) = inbox.remove(&inbox_cursor) {
-                    cloned_ctx.try_log(|logger| {
-                        info!(
-                            logger,
-                            "Adding block #{inbox_cursor} to next sequence (# blocks inboxed: {})",
-                            inbox.len()
-                        )
-                    });
-                    chunk.push((block, compacted_block));
-                    inbox_cursor += 1;
-                }
-                if !chunk.is_empty() {
-                    if let Some(ref tx) = blocks_post_processor {
-                        let _ = tx.send(chunk);
-                    }
-                } else {
-                    if blocks_processed == number_of_blocks_to_process {
-                        cloned_ctx.try_log(|logger| {
-                            info!(
-                                logger,
-                                "Local block storage successfully seeded with #{blocks_processed} blocks"
-                            )
-                        });
-                        break;
-                    }
-                }
-                sleep(Duration::from_secs(3));
-            }
-            ()
-        })
-        .expect("unable to spawn thread");
-
-    let mut thread_index = 0;
-    while let Some(res) = set.join_next().await {
-        let block = res.unwrap().unwrap();
-
-        let _ = tx_thread_pool[thread_index].send(Some(block));
-        if let Some(block_height) = block_heights.pop_front() {
-            let config = moved_config.clone();
-            let ctx = ctx.clone();
-            let http_client = moved_http_client.clone();
-            set.spawn(try_download_block_bytes_with_retry(
-                http_client,
-                block_height,
-                config,
-                ctx,
-            ));
-        }
-        thread_index = (thread_index + 1) % hord_config.ingestion_thread_max;
-    }
-
-    ctx.try_log(|logger| info!(logger, "Gargbage collecting will start"));
-
-    for tx in tx_thread_pool.iter() {
-        let _ = tx.send(None);
-    }
-
-    for handle in thread_pool_handles.into_iter() {
-        let _ = handle.join();
-    }
-
-    let _ = storage_thread.join();
-    let _ = set.shutdown();
-
-    ctx.try_log(|logger| info!(logger, "Gargbage collecting did finish"));
-
-    // match guard.report().build() {
-    //     Ok(report) => {
-    //         ctx.try_log(|logger| {
-    //             slog::info!(logger, "Generating report");
-    //         });
-    //         let file = std::fs::File::create("hord-perf.svg").unwrap();
-    //         report.flamegraph(file).unwrap();
-    //     }
-    //     Err(e) => {
-    //         ctx.try_log(|logger| {
-    //             slog::error!(logger, "Reporting failed: {}", e.to_string());
-    //         });
-    //     }
-    // }
-
-    Ok(())
 }
