+++ components/hord-cli/src/hord/pipeline/mod.rs
@@ -1,0 +1,222 @@
+use chainhook_sdk::observer::BitcoinConfig;
+use chainhook_sdk::types::BitcoinBlockData;
+use chainhook_sdk::utils::Context;
+use crossbeam_channel::bounded;
+use std::collections::{HashMap, VecDeque};
+use std::thread::sleep;
+use std::time::Duration;
+use tokio::task::JoinSet;
+
+use crate::config::Config;
+use crate::db::LazyBlock;
+
+use chainhook_sdk::indexer::bitcoin::{
+    build_http_client, parse_downloaded_block, try_download_block_bytes_with_retry,
+};
+
+use super::parse_ordinals_and_standardize_block;
+
+pub async fn download_and_pipeline_blocks(
+    config: &Config,
+    start_block: u64,
+    end_block: u64,
+    start_sequencing_blocks_at_height: u64,
+    blocks_post_processor: Option<crossbeam_channel::Sender<Vec<(BitcoinBlockData, LazyBlock)>>>,
+    ctx: &Context,
+) -> Result<(), String> {
+    // let guard = pprof::ProfilerGuardBuilder::default()
+    //     .frequency(20)
+    //     .blocklist(&["libc", "libgcc", "pthread", "vdso"])
+    //     .build()
+    //     .unwrap();
+
+    let bitcoin_config = BitcoinConfig {
+        username: config.network.bitcoind_rpc_username.clone(),
+        password: config.network.bitcoind_rpc_password.clone(),
+        rpc_url: config.network.bitcoind_rpc_url.clone(),
+        network: config.network.bitcoin_network.clone(),
+        bitcoin_block_signaling: config.network.bitcoin_block_signaling.clone(),
+    };
+
+    let hord_config = config.get_hord_config();
+
+    let number_of_blocks_to_process = end_block - start_block + 1;
+
+    let (block_compressed_tx, block_compressed_rx) = crossbeam_channel::bounded(50);
+    let http_client = build_http_client();
+
+    let moved_config = bitcoin_config.clone();
+    let moved_ctx = ctx.clone();
+    let moved_http_client = http_client.clone();
+
+    let mut set = JoinSet::new();
+
+    let mut block_heights = VecDeque::from((start_block..=end_block).collect::<Vec<u64>>());
+
+    for _ in 0..4 {
+        if let Some(block_height) = block_heights.pop_front() {
+            let config = moved_config.clone();
+            let ctx = moved_ctx.clone();
+            let http_client = moved_http_client.clone();
+            sleep(Duration::from_millis(500));
+            set.spawn(try_download_block_bytes_with_retry(
+                http_client,
+                block_height,
+                config,
+                ctx,
+            ));
+        }
+    }
+
+    let moved_ctx: Context = ctx.clone();
+    let moved_bitcoin_network = bitcoin_config.network.clone();
+
+    let mut tx_thread_pool = vec![];
+    let mut rx_thread_pool = vec![];
+    let mut thread_pool_handles = vec![];
+
+    for _ in 0..hord_config.ingestion_thread_max {
+        let (tx, rx) = bounded::<Option<Vec<u8>>>(hord_config.ingestion_thread_queue_size);
+        tx_thread_pool.push(tx);
+        rx_thread_pool.push(rx);
+    }
+
+    for rx in rx_thread_pool.into_iter() {
+        let block_compressed_tx_moved = block_compressed_tx.clone();
+        let moved_ctx: Context = moved_ctx.clone();
+        let moved_bitcoin_network = moved_bitcoin_network.clone();
+
+        let handle = hiro_system_kit::thread_named("Block data compression")
+            .spawn(move || {
+                while let Ok(Some(block_bytes)) = rx.recv() {
+                    let raw_block_data =
+                        parse_downloaded_block(block_bytes).expect("unable to parse block");
+                    let compressed_block = LazyBlock::from_full_block(&raw_block_data)
+                        .expect("unable to compress block");
+                    let block_data = parse_ordinals_and_standardize_block(
+                        raw_block_data,
+                        &moved_bitcoin_network,
+                        &moved_ctx,
+                    )
+                    .expect("unable to deserialize block");
+
+                    let _ = block_compressed_tx_moved.send(Some((block_data, compressed_block)));
+                }
+            })
+            .expect("unable to spawn thread");
+        thread_pool_handles.push(handle);
+    }
+
+    let cloned_ctx = ctx.clone();
+
+    let storage_thread = hiro_system_kit::thread_named("Ordered blocks dispatcher")
+        .spawn(move || {
+            let mut inbox = HashMap::new();
+            let mut inbox_cursor = start_sequencing_blocks_at_height.max(start_block);
+            let mut blocks_processed = 0;
+
+            loop {
+                // Dequeue all the blocks available
+                let mut new_blocks = vec![];
+                while let Ok(Some((block, compacted_block))) = block_compressed_rx.try_recv()
+                {
+                    blocks_processed += 1;
+                    new_blocks.push((block, compacted_block))
+                }
+                // 
+                let mut ooo_processing = vec![];
+                for (block, compacted_block) in new_blocks.into_iter() {
+                    let block_index = block.block_identifier.index;
+                    if block_index >= start_sequencing_blocks_at_height {
+                        inbox.insert(block_index, (block, compacted_block));
+                    } else {
+                        ooo_processing.push((block, compacted_block));
+                        // todo: do something
+                    }
+                }
+
+                // In order processing: construct the longest sequence of known blocks
+                let mut chunk = Vec::new();
+                while let Some((block, compacted_block)) = inbox.remove(&inbox_cursor) {
+                    cloned_ctx.try_log(|logger| {
+                        info!(
+                            logger,
+                            "Adding block #{inbox_cursor} to next sequence (# blocks inboxed: {})",
+                            inbox.len()
+                        )
+                    });
+                    chunk.push((block, compacted_block));
+                    inbox_cursor += 1;
+                }
+                if !chunk.is_empty() {
+                    if let Some(ref tx) = blocks_post_processor {
+                        let _ = tx.send(chunk);
+                    }
+                } else {
+                    if blocks_processed == number_of_blocks_to_process {
+                        cloned_ctx.try_log(|logger| {
+                            info!(
+                                logger,
+                                "Local block storage successfully seeded with #{blocks_processed} blocks"
+                            )
+                        });
+                        break;
+                    }
+                }
+                sleep(Duration::from_secs(3));
+            }
+            ()
+        })
+        .expect("unable to spawn thread");
+
+    let mut thread_index = 0;
+    while let Some(res) = set.join_next().await {
+        let block = res.unwrap().unwrap();
+
+        let _ = tx_thread_pool[thread_index].send(Some(block));
+        if let Some(block_height) = block_heights.pop_front() {
+            let config = moved_config.clone();
+            let ctx = ctx.clone();
+            let http_client = moved_http_client.clone();
+            set.spawn(try_download_block_bytes_with_retry(
+                http_client,
+                block_height,
+                config,
+                ctx,
+            ));
+        }
+        thread_index = (thread_index + 1) % hord_config.ingestion_thread_max;
+    }
+
+    ctx.try_log(|logger| info!(logger, "Gargbage collecting will start"));
+
+    for tx in tx_thread_pool.iter() {
+        let _ = tx.send(None);
+    }
+
+    for handle in thread_pool_handles.into_iter() {
+        let _ = handle.join();
+    }
+
+    let _ = storage_thread.join();
+    let _ = set.shutdown();
+
+    ctx.try_log(|logger| info!(logger, "Gargbage collecting did finish"));
+
+    // match guard.report().build() {
+    //     Ok(report) => {
+    //         ctx.try_log(|logger| {
+    //             slog::info!(logger, "Generating report");
+    //         });
+    //         let file = std::fs::File::create("hord-perf.svg").unwrap();
+    //         report.flamegraph(file).unwrap();
+    //     }
+    //     Err(e) => {
+    //         ctx.try_log(|logger| {
+    //             slog::error!(logger, "Reporting failed: {}", e.to_string());
+    //         });
+    //     }
+    // }
+
+    Ok(())
+}
