--- components/hord-cli/src/hord/tests/mod.rs
@@ -1,1 +1,0 @@
-
